L9/pyproject.toml
[project]
name = "rssmds"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "pyyaml",
    "beautifulsoup4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "types-requests",
    "types-PyYAML",
    "types-beautifulsoup4",
]

[tool.mypy]
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
